# perf(duckdb): speed up performance with wide tables