@cpcloud 2023-08-10
There's a strange bug (https://github.com/duckdb/duckdb/issues/8539) that prevents us from using the faster `describe t` version of things for now :(