@cpcloud 2023-08-11
Alright, I was able to workaround the upstream bug using the `information_schema.columns` view to get the view metadata much faster than `DESCRIBE SELECT * FROM t`, along with the correct ordinal position of the columns.

Here's the current profile:

**`master`**:

![image](https://github.com/ibis-project/ibis/assets/417981/06278899-6d8d-420c-ae30-6045836c48a8)

**This PR**:

![image](https://github.com/ibis-project/ibis/assets/417981/5f803207-3947-4bc0-9316-027b90ed8746)

So about a 2x speed up overall, with most the improvement coming from querying static type information instead of using `DESCRIBE SELECT *`